Extract shared web session plumbing into @tinyburg/web-auth
The residual code both apps carried byte-for-byte identical copies of
after the RelyingParty ports: the sha256/randomSecret crypto helpers,
the CookiePolicy service (NODE_ENV-driven Secure and __Host- prefix
policy), and the isLocalPath/returnToParam open redirect guard. Each
app keeps its own cookie names; only the mechanism moves. Both apps'
cookies.ts re-export CookiePolicy so call sites are untouched.
