+++ apps/authproxy/package.json
@@ -16,6 +16,7 @@
         "@tinyburg/nimblebit-sdk": "workspace:*",
         "@tinyburg/tinytower-sdk": "workspace:*",
         "@tinyburg/ui": "workspace:*",
+        "@tinyburg/web-auth": "workspace:*",
         "effect": "4.0.0-beta.105",
         "effect-oidc": "0.0.8",
         "foldkit": "0.140.1",
