+++ apps/tinyburg.app/server/cookies.ts
@@ -1,36 +1,19 @@
 import type { SqlError } from "effect/unstable/sql";
 
-import { Config, Context, Effect, Layer, Option, type Schema } from "effect";
+import { Effect, Option, type Schema } from "effect";
 import { HttpServerRequest } from "effect/unstable/http";
 
 import type { Session, User } from "../domain/models.ts";
 
+import { CookiePolicy } from "@tinyburg/web-auth/CookiePolicy";
+import { sha256 } from "@tinyburg/web-auth/Crypto";
+
 import { SessionsRepository } from "../domain/sessions.ts";
-import { sha256 } from "./crypto.ts";
+
+export { CookiePolicy };
 
 /** The name of the cookie used to store the provider session. */
 export const PROVIDER_SESSION_COOKIE_NAME = "tinyburg_provider_session";
-
-/**
- * The site's cookie policy, resolved once at boot so request handlers never
- * carry a config failure channel.
- *
- * Cookies demand https unless the environment explicitly says development, so
- * a production deploy that forgets to set NODE_ENV still ships Secure
- * cookies. Secure cookies also carry the __Host- prefix, which browsers only
- * accept over https, from the exact host, with no Domain attribute: a hostile
- * subdomain can then never plant a copy of the session or OAuth round-trip
- * cookies. Plain names in development, where http would refuse the prefix.
- */
-export class CookiePolicy extends Context.Service<CookiePolicy>()("@tinyburg/tinyburg.app/server/CookiePolicy", {
-    make: Effect.map(Config.string("NODE_ENV").pipe(Config.withDefault("production")), (env) => {
-        const secure = env !== "development";
-        const name = (base: string): string => (secure ? `__Host-${base}` : base);
-        return { secure, name } as const;
-    }),
-}) {
-    static readonly Default = Layer.effect(this, CookiePolicy.make);
-}
 
 /**
  * The user and session riding the provider session cookie, if any. None for a
