+++ apps/tinyburg.app/tsconfig.json
@@ -2,7 +2,8 @@
     "extends": "../../tsconfig.base.jsonc",
     "references": [
         { "path": "../../packages/nimblebit-sdk/tsconfig.src.json" },
-        { "path": "../../packages/ui/tsconfig.src.json" }
+        { "path": "../../packages/ui/tsconfig.src.json" },
+        { "path": "../../packages/web-auth/tsconfig.src.json" }
     ],
     "include": [
         "client",
