+++ packages/web-auth/src/Crypto.ts
@@ -1,11 +1,20 @@
+/**
+ * The two primitives every session and OAuth flow leans on: hashing secrets
+ * and minting them.
+ *
+ * @since 1.0.0
+ */
+
 import { Effect, Encoding } from "effect";
 
 import * as crypto from "node:crypto";
 
 /**
  * SHA-256 as base64url, the shape OAuth uses for the PKCE S256 challenge and
- * for everything we keep a hash of rather than the secret itself: session
- * cookies alike.
+ * for everything we keep a hash of rather than the secret itself: client
+ * secrets, authorization codes and session cookies alike.
+ *
+ * @since 1.0.0
  */
 export const sha256 = (value: string): Effect.Effect<string> =>
     Effect.map(
@@ -16,6 +25,8 @@ export const sha256 = (value: string): Effect.Effect<string> =>
 /**
  * A 384-bit random string. Long enough that guessing one is not a strategy,
  * which is what OAuth state, PKCE verifiers and session cookies all need.
+ *
+ * @since 1.0.0
  */
 export const randomSecret = (): string =>
     Array.from(crypto.getRandomValues(new Uint8Array(48)), (byte) => byte.toString(16).padStart(2, "0")).join("");
