+++ packages/web-auth/src/index.ts
@@ -1,0 +1,26 @@
+/**
+ * @since 1.0.0
+ */
+
+/**
+ * The one cookie policy every tinyburg web app applies to every cookie it
+ * sets.
+ *
+ * @since 1.0.0
+ */
+export * as CookiePolicy from "./CookiePolicy.ts";
+
+/**
+ * The two primitives every session and OAuth flow leans on: hashing secrets
+ * and minting them.
+ *
+ * @since 1.0.0
+ */
+export * as Crypto from "./Crypto.ts";
+
+/**
+ * The open-redirect guard for `returnTo`-style parameters.
+ *
+ * @since 1.0.0
+ */
+export * as Redirect from "./Redirect.ts";
