+++ packages/web-auth/tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "./tsconfig.src.json",
+    "compilerOptions": {
+        "tsBuildInfoFile": ".tsbuildinfo/build.tsbuildinfo",
+        "outDir": "dist",
+        "stripInternal": true
+    }
+}
