+++ packages/web-auth/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "../../tsconfig.base.jsonc",
+    "include": [],
+    "references": [{ "path": "tsconfig.src.json" }]
+}
