+++ packages/web-auth/tsconfig.src.json
@@ -1,0 +1,11 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "../../tsconfig.base.jsonc",
+    "include": ["src"],
+    "compilerOptions": {
+        "tsBuildInfoFile": ".tsbuildinfo/src.tsbuildinfo",
+        "outDir": ".tsbuildinfo/src",
+        "rootDir": "src",
+        "types": ["node"]
+    }
+}
