+++ tsconfig.build.json
@@ -13,6 +13,7 @@
         { "path": "packages/tinytower-sdk/tsconfig.build.json" },
         { "path": "packages/trading-sdk/tsconfig.build.json" },
         { "path": "packages/ui/tsconfig.build.json" },
+        { "path": "packages/web-auth/tsconfig.build.json" },
         { "path": "apps/archivist/tsconfig.json" },
         { "path": "apps/authproxy/tsconfig.json" },
         { "path": "apps/auto-gold-bits/tsconfig.json" },
