+++ tsconfig.json
@@ -10,6 +10,7 @@
         { "path": "packages/tinytower-sdk" },
         { "path": "packages/trading-sdk" },
         { "path": "packages/ui" },
+        { "path": "packages/web-auth" },
         { "path": "apps/archivist" },
         { "path": "apps/authproxy" },
         { "path": "apps/auto-gold-bits" },
